refactor(generate): streamline generate logic and enhance type definitions
