+++ src/app/[locale]/topics/[topicId]/(topic)/context/GenerateContext.tsx
@@ -1,11 +1,5 @@
+import { ICommonGenerateOptions } from '@/hooks/generate/type';
 import React, { createContext, useContext, useState, ReactNode } from 'react';
-
-export interface ICommonGenerateOptions {
-    numberOfItem: number;
-    difficulty: string;
-    focus: string;
-    listType: string[];
-}
 
 interface IGenerateContext {
     options: ICommonGenerateOptions;
