+++ src/hooks/generate/useGenerate.tsx
@@ -6,36 +6,10 @@ import { ApiResponsePubGenContent, ISseData } from '@/app/[locale]/generate';
 import { BASE_URL_STREAM_GENERATE, URL_API_GENERATE } from '@/app/[locale]/generate/utils/constant';
 import { compressContent } from '@/app/[locale]/generate/helper/compress';
 import { toast } from '../use-toast';
-import { NodesData } from '@/app/[locale]/topics/[topicId]/(topic)/types/generate.type';
-import { ICommonGenerateOptions } from '@/app/[locale]/topics/[topicId]/(topic)/context/GenerateContext';
 import { safeDestructure } from '@/utils';
+import { IGenerateRequest, ValidateGeneratedDataFn } from './type';
 
-export interface IGenerateOptions {
-    commonGenerateOptions?: ICommonGenerateOptions;
-    nodesData?: NodesData;
-}
-
-export interface IGenerateRequest {
-    content: string;
-    method: string;
-    type: string;
-    options?: IGenerateOptions;
-}
-
-interface ValidateDataSuccess<TRes> {
-    ok: true;
-    data: TRes;
-}
-
-interface ValidateDataError<TRes> {
-    ok: false;
-    error: unknown;
-    data?: TRes;
-}
-
-export type ValidateGeneratedDataResult<TRes> = ValidateDataSuccess<TRes> | ValidateDataError<TRes>;
-export type ValidateGeneratedDataFn<TRes> = (data: TRes) => ValidateGeneratedDataResult<TRes>;
-
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 export interface UsePostOptions<TReq, TRes> {
     validateGeneratedData?: ValidateGeneratedDataFn<TRes>;
     onSuccess?: (data: TRes) => void;
@@ -155,7 +129,7 @@ export default function useGenerate<TRes = unknown>(options?: UsePostOptions<IGe
                 if (!validateResult.ok) {
                     onError?.(validateResult.error);
                     return;
-                } 
+                }
                 dataValidated = validateResult.data;
             }
